refactor: remove refit references
