--- Source/Iwentys.Bot/ApiSdk/IAssignmentApi.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using Iwentys.Models.Transferable;
-using Refit;
-
-namespace Iwentys.ClientBot.ApiSdk
-{
-    public interface IAssignmentApi
-    {
-        [Get("/api/Assignment")]
-        Task<List<AssignmentInfoDto>> Get();
-
-        [Post("/api/Assignment")]
-        Task<List<AssignmentInfoDto>> Create([Body] AssignmentCreateDto assignmentCreateDto);
-    }
-}
