--- Source/Iwentys.Bot/ApiSdk/IStudyLeaderboardApi.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using Iwentys.Models.Transferable.Study;
-using Iwentys.Models.Types;
-using Refit;
-
-namespace Iwentys.ClientBot.ApiSdk
-{
-    public interface IStudyLeaderboardApi
-    {
-        [Get("/api/StudyLeaderboard")]
-        public Task<List<StudyLeaderboardRow>> GetStudentsRating(int? subjectId, int? courseId, int? groupId, StudySemester? semester);
-    }
-}
