+++ Source/Iwentys.Bot/Commands/Guild/GetGuildsCommand.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using FluentResults;
-using Iwentys.ClientBot.ApiSdk;
+using Iwentys.ClientBot.Tools;
 using Iwentys.Models.Tools;
 using Iwentys.Models.Transferable.Guilds;
 using Tef.BotFramework.Abstractions;
