+++ Source/Iwentys.Bot/Commands/Student/GetCurrentStudentCommand.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using FluentResults;
-using Iwentys.ClientBot.ApiSdk;
+using Iwentys.ApiClient.OpenAPIService;
 using Iwentys.ClientBot.Tools;
 using Iwentys.Core.DomainModel;
 using Iwentys.Models.Transferable.Students;
@@ -33,8 +33,8 @@ namespace Iwentys.ClientBot.Commands.Student
         public async Task<Result<string>> ExecuteAsync(CommandArgumentContainer args)
         {
             AuthorizedUser currentUser = _userIdentifier.GetUser(args.Sender.UserSenderId);
-            IwentysApiProvider userProvider = await _userIdentifier.GetProvider(args.Sender.UserSenderId, _api).ConfigureAwait(false);
-            StudentFullProfileDto profile = await userProvider.Client.ApiStudentGetAsync(currentUser.Id).ConfigureAwait(false);
+            Client client = await _userIdentifier.GetProvider(args.Sender.UserSenderId, _api).ConfigureAwait(false);
+            StudentFullProfileDto profile = await client.ApiStudentGetAsync(currentUser.Id).ConfigureAwait(false);
             return Result.Ok(profile.FormatFullInfo());
         }
