+++ Source/Iwentys.Bot/Commands/StudentLeaderboard/GetStudentsRatingCommand.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 using FluentResults;
-using Iwentys.ClientBot.ApiSdk;
+using Iwentys.ClientBot.Tools;
 using Iwentys.Models.Tools;
 using Iwentys.Models.Transferable.Study;
 using Tef.BotFramework.Abstractions;
