+++ Source/Iwentys.Bot/Tools/UserIdentifier.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using Iwentys.ClientBot.ApiSdk;
+using Iwentys.ApiClient.OpenAPIService;
 using Iwentys.Core.DomainModel;
 
 namespace Iwentys.ClientBot.Tools
@@ -8,11 +8,11 @@ namespace Iwentys.ClientBot.Tools
     public class UserIdentifier
     {
         private readonly Dictionary<int, int> _usersByTelegramId;
-        private readonly Dictionary<int, IwentysApiProvider> _userProviders;
+        private readonly Dictionary<int, Client> _userProviders;
 
         public UserIdentifier()
         {
-            _userProviders = new Dictionary<int, IwentysApiProvider>();
+            _userProviders = new Dictionary<int, Client>();
             _usersByTelegramId = new Dictionary<int, int>();
         }
 
@@ -28,13 +28,13 @@ namespace Iwentys.ClientBot.Tools
                 : 228617);
         }
 
-        public async Task<IwentysApiProvider> GetProvider(int messengerUserId, IwentysApiProvider apiProvider)
+        public async Task<Client> GetProvider(int messengerUserId, IwentysApiProvider apiProvider)
         {
-            var user = GetUser(messengerUserId);
+            AuthorizedUser user = GetUser(messengerUserId);
             if (!_userProviders.ContainsKey(user.Id))
             {
                 var token = await apiProvider.Client.ApiDebugcommandLoginorcreateAsync(user.Id).ConfigureAwait(false);
-                _userProviders[user.Id] = new IwentysApiProvider(token.Token);
+                _userProviders[user.Id] = IwentysApiProvider.Create(token.Token);
             }
 
             return _userProviders[user.Id];
